script: Move num module to script_bindings.
