+++ components/script/dom/bindings/mod.rs
@@ -151,7 +151,6 @@ pub(crate) mod interface;
 pub(crate) mod iterable;
 pub(crate) mod like;
 pub(crate) mod namespace;
-pub(crate) mod num;
 pub(crate) mod principals;
 pub(crate) mod proxyhandler;
 pub(crate) mod refcounted;
@@ -166,6 +165,8 @@ pub(crate) mod transferable;
 pub(crate) mod utils;
 pub(crate) mod weakref;
 pub(crate) mod xmlname;
+
+pub(crate) use script_bindings::num;
 
 /// Generated JS-Rust bindings.
 #[allow(missing_docs, non_snake_case)]
